math_parser: move diag_value to a separate unit
